+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/symmetric/KeyGen.kt
@@ -5,12 +5,11 @@ import at.asitplus.catching
 import at.asitplus.signum.HazardousMaterials
 import at.asitplus.signum.indispensable.asn1.encoding.bitLength
 import at.asitplus.signum.indispensable.misc.BitLength
-import at.asitplus.signum.SecureRandom
+import org.kotlincrypto.random.CryptoRand
 import kotlin.jvm.JvmName
-import kotlin.random.Random
 
-private inline fun randomBytes(n: Int, random: Random = SecureRandom): ByteArray =
-    random.nextBytes(n)
+private inline fun randomBytes(n: Int, random: CryptoRand = CryptoRand.Default): ByteArray =
+    ByteArray(n).also {   random.nextBytes(it)}
 
 /**
  * Generates a fresh random key for this algorithm.
@@ -18,7 +17,7 @@ private inline fun randomBytes(n: Int, random: Random = SecureRandom): ByteArray
 @Suppress("INVISIBLE_MEMBER", "INVISIBLE_REFERENCE", "UNCHECKED_CAST")
 @kotlin.internal.LowPriorityInOverloadResolution
 suspend fun <A : AuthCapability<out K>, I : NonceTrait, K : KeyType> SymmetricEncryptionAlgorithm<A, I, K>.randomKey() =
-    @OptIn(HazardousMaterials::class) randomKey(SecureRandom)
+    @OptIn(HazardousMaterials::class) randomKey(CryptoRand.Default)
 /**
  * Generates a fresh random key for this algorithm.
  */
@@ -26,7 +25,7 @@ suspend fun <A : AuthCapability<out K>, I : NonceTrait, K : KeyType> SymmetricEn
 @kotlin.internal.LowPriorityInOverloadResolution
 @HazardousMaterials("The default randomness source is a secure random. Override it for reproducible tests, but if you run out of entropy in production, find the root cause!")
 suspend fun <A : AuthCapability<out K>, I : NonceTrait, K : KeyType> SymmetricEncryptionAlgorithm<A, I, K>.randomKey(
-    random: Random
+    random: CryptoRand
 ): SymmetricKey<A, I, out K> =
     keyFromInternal(
         randomBytes(keySize.bytes.toInt(), random),
@@ -44,7 +43,7 @@ suspend fun <A : AuthCapability<out K>, I : NonceTrait, K : KeyType> SymmetricEn
 @Suppress("UNCHECKED_CAST")
 suspend fun <I : NonceTrait> SymmetricEncryptionAlgorithm<AuthCapability.Authenticated.WithDedicatedMac, I, KeyType.WithDedicatedMacKey>.randomKey(
     macKeyLength: BitLength
-) = @OptIn(HazardousMaterials::class) randomKey(macKeyLength, SecureRandom)
+) = @OptIn(HazardousMaterials::class) randomKey(macKeyLength, CryptoRand.Default)
 /**
  * Generates a fresh random key for this algorithm.
  * [macKeyLength] can be specified to override [preferredMacKeyLength].
@@ -54,7 +53,7 @@ suspend fun <I : NonceTrait> SymmetricEncryptionAlgorithm<AuthCapability.Authent
 @HazardousMaterials("The default randomness source is a secure random. Override it for reproducible tests, but if you run out of entropy in production, find the root cause!")
 suspend fun <I : NonceTrait> SymmetricEncryptionAlgorithm<AuthCapability.Authenticated.WithDedicatedMac, I, KeyType.WithDedicatedMacKey>.randomKey(
     macKeyLength: BitLength,
-    random: Random
+    random: CryptoRand
 ): SymmetricKey.WithDedicatedMac<I> =
     keyFromInternal(
         randomBytes(keySize.bytes.toInt(), random),
@@ -68,7 +67,7 @@ suspend fun <I : NonceTrait> SymmetricEncryptionAlgorithm<AuthCapability.Authent
  */
 @HazardousMaterials("Don't explicitly generate nonces!")
 fun SymmetricEncryptionAlgorithm<*, NonceTrait.Required, *>.randomNonce(): ByteArray =
-    randomBytes((nonceSize.bytes).toInt(), SecureRandom)
+    randomBytes((nonceSize.bytes).toInt(), CryptoRand.Default)
 
 /**
  * Generates a new random nonce matching the Nonce size of this algorithm.
@@ -76,7 +75,7 @@ fun SymmetricEncryptionAlgorithm<*, NonceTrait.Required, *>.randomNonce(): ByteA
  */
 @HazardousMaterials("Don't explicitly generate nonces!")
 @HazardousMaterials("The default randomness source is a secure random. Override it for reproducible tests, but if you run out of entropy in production, find the root cause!")
-fun SymmetricEncryptionAlgorithm<*, NonceTrait.Required, *>.randomNonce(random: Random): ByteArray =
+fun SymmetricEncryptionAlgorithm<*, NonceTrait.Required, *>.randomNonce(random: CryptoRand): ByteArray =
     randomBytes((nonceSize.bytes).toInt(), random)
 
 
@@ -154,8 +153,8 @@ fun <A : AuthCapability<KeyType.WithDedicatedMacKey>, I : NonceTrait> SymmetricE
 
 
 suspend fun SpecializedSymmetricEncryptionAlgorithm.randomKey() =
-    @OptIn(HazardousMaterials::class) randomKey(SecureRandom)
+    @OptIn(HazardousMaterials::class) randomKey(CryptoRand.Default)
 
 @HazardousMaterials("The default randomness source is a secure random. Override it for reproducible tests, but if you run out of entropy in production, find the root cause!")
-suspend fun SpecializedSymmetricEncryptionAlgorithm.randomKey(random: Random) =
+suspend fun SpecializedSymmetricEncryptionAlgorithm.randomKey(random: CryptoRand) =
     algorithm.randomKey(random)